- renamed screen module for consistency and to reduce redundancy on imports - renamed Sound screen to Audio for consistency - made Screen API slightly more pythonic. Similar plans for Widget
