+++ usf/gui.py
@@ -30,18 +30,18 @@ from pygame.locals import QUIT
 # Our modules
 from usf.font import fonts
 from usf.game import Game
-from usf.screen.about import About
-from usf.screen.characters import Characters
-from usf.screen.configure import Configure
-from usf.screen.keyboard import Keyboard
-from usf.screen.level import Level
-from usf.screen.main_screen import MainScreen
-from usf.screen.resume import Resume
-from usf.screen.display import Display
-from usf.screen.sound import Sound
-from usf.screen.network_screen import NetworkScreen
-from usf.screen.network_join_screen import NetworkJoinScreen
-from usf.screen.network_game_conf_screen import NetworkGameConfScreen
+from usf.screens.about import About
+from usf.screens.characters import Characters
+from usf.screens.configure import Configure
+from usf.screens.keyboard import Keyboard
+from usf.screens.level import Level
+from usf.screens.main_screen import MainScreen
+from usf.screens.resume import Resume
+from usf.screens.display import Display
+from usf.screens.sound import Sound
+from usf.screens.network_screen import NetworkScreen
+from usf.screens.network_join_screen import NetworkJoinScreen
+from usf.screens.network_game_conf_screen import NetworkGameConfScreen
 from usf.skin import Skin
 from usf.translation import _
 from usf.widgets.widget import optimize_size
