+++ usf/screens/about.py
@@ -1,5 +1,5 @@
 ################################################################################
-# copyright 2010 Gabriel Pettier <gabriel.pettier@gmail.com>                   #
+# copyright 2010 Lucas Baudin <xapantu@gmail.com>                              #
 #                                                                              #
 # This file is part of Ultimate Smash Friends.                                 #
 #                                                                              #
@@ -18,29 +18,33 @@
 ################################################################################
 
 '''
-The Pause/Resume screen, allow to go to configuration, to leave game or to
-resume playing the current game.
+The About/Credit screen.
 
 '''
 
-from usf.screen.screen import Screen
-from usf.widgets.button import Button
+from usf.screens.screen import Screen
+from usf import loaders
+
 from usf.widgets.box import VBox
+from usf.widgets.button import Button
+from usf.widgets.paragraph import Paragraph
+
 from usf.translation import _
 
+class About(Screen):
 
-class Resume(Screen):
     def init(self):
+        self.name = (_("about"))
         self.add(VBox())
-        self.widget.add(Button('Resume'))
-        self.widget.add(Button(_('Configure')))
-        self.widget.add(Button('Quit the game'))
+
+        self.widget.add(Paragraph('CREDITS'),
+                size=(
+                    490*loaders.get_config().general.WIDTH/800,
+                    300*loaders.get_config().general.HEIGHT/600))
+
+        self.widget.add(Button(_('Back')),
+                        margin=55)
 
     def callback(self, action):
-        if action.text == 'Resume':
-            return "game:continue"
-        if action.text == 'Quit the game':
-            return 'game:stop'
-        if action.text == _('Configure'):
-            return 'goto:configure'
-
+        if action.text == _('Back'):
+            return "goto:back"
