+++ usf/screens/characters.py
@@ -18,15 +18,11 @@
 ################################################################################
 
 '''
-Show connected users list, player choices and level choice for a network game,
-this screen should be seen identically by the player hosting the game and the
-player(s) joining it, only each player can only chose his character. Level can
-be chosen by both, the game start when both player have clicked the "start
-game" button (maybe with some time delay, and with possibility to cancel?).
+Describe the characters choice UI (for starting a game)
 
 '''
 
-from usf.screen.screen import Screen
+from usf.screens.screen import Screen
 from usf import loaders
 CONFIG = loaders.get_config()
 
@@ -36,19 +32,16 @@ from usf.widgets.image import Image
 from usf.widgets.spinner import Spinner
 from usf.widgets.label import Label
 from usf.widgets.checkbox_text import TextCheckBox
-from usf.widgets.text_entry import TextEntry
-from usf.widgets.coverflow import Coverflow
 
 from usf.translation import _
 
 from usf import entity_skin
 import os
 from os.path import join
-import logging
 
 
-class NetworkGameConfScreen(Screen):
-    """ the network game configuration screen
+class Characters(Screen):
+    """ the character selection screen
     """
     name_pl1 = 0
     name_pl2 = 0
@@ -61,7 +54,7 @@ class NetworkGameConfScreen(Screen):
         self.game_data['character_file'] = []
 
         #I18N: The title of the screen where players can choose their character.
-        self.set_name(_("characters"))
+        self.name = _("characters")
 
         #create a character file to have the [?] image
         self.game_data['character_file'].append(join('characters', 'none'))
@@ -73,48 +66,6 @@ class NetworkGameConfScreen(Screen):
         files = os.listdir(join(CONFIG.system_path, 'characters'))
         files.sort()
 
-        self.load_chararacters(files)
-
-        self.add(VBox())
-
-        self.portrait = Image(join(self.game_data['character_file'][0],
-            "portrait.png"))
-
-        self.player_spinner = Spinner(self.character)
-        player_vbox = VBox()
-        player_vbox.add(Label(_('Player name'), align='center'))
-        player_vbox.add(TextEntry(_('unnamed player')))
-        player_vbox.add(Spinner(self.character))
-        player_vbox.add(self.portrait, margin_left=65, margin=5, size=(50, 50))
-
-        hbox = HBox()
-        # adding the two box which contains the spinner and the name of the
-        # characters
-        hbox.add(player_vbox, margin=400)
-        self.widget.add(hbox, margin=150)
-
-        #create a level image for every directory in the level directory.
-        files = os.listdir(os.path.join( CONFIG.system_path, 'levels'))
-        files.sort()
-
-        coverflow_data = self.load_levels(files)
-
-        self.coverflow = Coverflow(coverflow_data)
-        self.widget.add(self.coverflow, size=(800, 275))
-
-        #next button to go to the level screen
-        self.widget.add(Button(_("Start")),
-            margin=83,
-            align="center")
-
-        #back button to come back to main screen
-        self.widget.add(Button(_('Back')),
-            margin=20,
-            align="center")
-
-    def load_chararacters(self, files):
-        """ append characters to self.character, loaded from files in "files".
-        """
         for f in files:
             try:
                 if f != "none":
@@ -132,38 +83,63 @@ class NetworkGameConfScreen(Screen):
             except IOError, e:
                 pass
 
-    def load_levels(self, files):
-        """ append a level in self.levels
-        """
-        coverflow_data = []
-        for f in files:
-            try:
-                if 'level.xml' in os.listdir(
-                        os.path.join(CONFIG.system_path, "levels", f)):
-                    coverflow_data.append([])
-                    coverflow_data[-1].append(f)
-                    coverflow_data[-1].append(
-                            join(
-                                CONFIG.system_path,
-                                "levels", f, "screenshot.png"))
-            except:
-                #XXX: catch all exceptions: BAD
-                logging.debug(str(f) +" is not a valid level.")
-        return coverflow_data
+        self.add(VBox())
+
+        self.checkboxes_ai = []
+        self.portraits = []
+        self.player_spinner = []
+        self.player_vbox = [VBox(), VBox(), VBox(), VBox()]
+
+        for i in range(0, 4):
+            #I18N: Artificial Intelligence
+            self.checkboxes_ai.append(TextCheckBox(_("AI:")))
+            self.portraits.append(Image(
+                    join(self.game_data['character_file'][0], "portrait.png")))
+
+            self.player_spinner.append(Spinner(self.character))
+            #I18N: %s is the player number, it can be Player 1, Player2...
+            self.player_vbox[i].add(Label(_("Player %s").replace(
+                "%s", str(i+1))))
+            self.player_vbox[i].add(self.player_spinner[-1])
+            self.player_vbox[i].add(self.portraits[-1],
+                margin_left=65,
+                margin=5,
+                size=(50, 50))
+            self.player_vbox[i].add(
+                    self.checkboxes_ai[-1],
+                    margin_left=(180 - self.checkboxes_ai[-1].width) / 2)
+
+
+        hbox = HBox()
+        # adding the two box which contains the spinner and the name of the
+        # characters
+        for vbox in self.player_vbox:
+            hbox.add(vbox, margin=20)
+        self.widget.add(hbox, margin=50)
+
+        #next button to go to the level screen
+        self.widget.add(Button(_("Next")),
+            margin=83,
+            align="center")
+
+        #back button to come back to main screen
+        self.widget.add(Button(_('Back')),
+            margin=20,
+            align="center")
 
     def callback(self, action):
-        if action is self.player_spinner :
+        if action in self.player_spinner :
             #get the index of the player
             player_number = self.player_spinner.index(action)
 
             self.players[player_number] = action.get_index()
             #change the portrait
             self.portraits[player_number].setImage(
-                    join(
+                        join(
                         self.game_data['character_file'][action.get_index()],
                         "portrait.png"))
 
-        if action.text == _("Start"):
+        if action.text == _("Next"):
             i  = 0
             for player in self.players:
                 if player != 0:
